test: helpers to build user and mock resolve them
A mock user can be built and we can provide overrides to set specific parts of the user. The prisma mock resolves the findUnique, update, and delete operations on a single user, where we would pass the overrides into that function rather than building our own user. We can also mock that a user was not found (null).