test: unit testing userService (#28)
* chore: install @types/jest to autoComplete with editor

* feat: mock prisma client with user model

* test: userService getUserById

- Return user given valid id
- throw error when user not found

* test: describe by method within outer describe

* test: helpers to build user and mock resolve them

A mock user can be built and we can provide overrides to set specific parts of the user. The prisma mock resolves the findUnique, update, and delete operations on a single user, where we would pass the overrides into that function rather than building our own user. We can also mock that a user was not found (null).

* feat: check if user exists before delete

Attempting to delete a user will throw a prisma error to the controller that calls this function. To control this, we check if the user exists, and throw a specific error message if they do not.

* test: getAllUsers and deleteUsers testing on found and not found cases

* feat: check if user exists before updating them

If a user doesn't exist on a prisma update request, an error will automatically throw. To control this, we can check if the user exists first, or throw a custom error message if not found.

* test: updateUser testing for found and not found branches